requirements.txt
# ============================================
# Python Dependencies for Asistente DB
# ============================================

# Database
sqlalchemy>=2.0.0,<3.0.0
psycopg2-binary>=2.9.9
alembic>=1.13.0

# Environment
python-dotenv>=1.0.0

# Utilities (optional but recommended)
pydantic>=2.5.0  # For data validation
python-dateutil>=2.8.2

# Development tools (optional)
ipython>=8.18.0  # Better Python shell
black>=23.12.0  # Code formatter
flake8>=7.0.0  # Linter
pytest>=7.4.0  # Testing
pytest-cov>=4.1.0  # Test coverage
